backend/requirements.txt
pandas>=2.2
numpy>=1.26
pyyaml>=6.0
python-dateutil>=2.9
pyarrow>=16.0
duckdb>=1.1.0
yfinance>=0.2.43
requests>=2.32
tqdm>=4.66
